requirements.txt
anyio==4.0.0
appdirs==1.4.4
asgiref==3.7.2
click==8.1.7
h11==0.14.0
htmltools==0.3.0
idna==3.4
linkify-it-py==2.0.2
lxml==4.9.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
numpy==1.26.0
packaging==23.2
Pillow==9.5.0
python-multipart==0.0.6
python-pptx==0.6.22
shiny==0.5.1
sniffio==1.3.0
starlette==0.31.1
tk==0.1.0
typing_extensions==4.8.0
uc-micro-py==1.0.2
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
XlsxWriter==3.1.6
pandas==2.1.1
